#Setup base JWT configuration - #Handle API response with ResponseEntity- #Format response before sending by using global with Controller advice

+++ src/main/java/AsukaSan/jobLancer/controller/UserController.java
@@ -2,6 +2,9 @@ package AsukaSan.jobLancer.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,37 +15,46 @@ import org.springframework.web.bind.annotation.RestController;
 
 import AsukaSan.jobLancer.domain.User;
 import AsukaSan.jobLancer.service.UserService;
+import AsukaSan.jobLancer.utils.error.IdInvalidException;
 
 @RestController
 public class UserController {
     private final UserService userService;
+    private final PasswordEncoder passwordEncoder;
 
-    public UserController(UserService userService){
+    public UserController(UserService userService,  PasswordEncoder passwordEncoder){
         this.userService = userService;
+        this.passwordEncoder = passwordEncoder;
     }
 
-    @PostMapping("/user/create")
-    public User createNewUser(@RequestBody User userFromPostMan){
-        
+    @PostMapping("/users/create")
+    public ResponseEntity<User> createNewUser(@RequestBody User userFromPostMan){
+        String hashPassword = this.passwordEncoder.encode(userFromPostMan.getPassWord());
+        userFromPostMan.setPassWord(hashPassword);
         User accUser = this.userService.handleCreateUser(userFromPostMan);
-        return accUser;
+        return ResponseEntity.status(HttpStatus.CREATED).body(accUser);
     }
-    @GetMapping("/user/{id}")
-    public User getUserInfo(@PathVariable("id") long id){
-        return this.userService.fetchUserById(id);
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> getUserInfo(@PathVariable("id") long id){
+        User fetchUser = this.userService.fetchUserById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(fetchUser);
     }
     //delete user
-    @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable("id") long id){
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) throws IdInvalidException{
+        if(id > 1300){
+            throw new IdInvalidException("Id limited at 1000");
+        }
         this.userService.deleteUserById(id);
-        return "Delete completely";
+        return ResponseEntity.ok("Delete successfully");
     }
-    @GetMapping("/user")
-    public List<User> getAllUsersInfo(){
-        return this.userService.fetchAllUsers();
+    @GetMapping("/users")
+    public ResponseEntity<List<User>> getAllUsersInfo(){
+        return ResponseEntity.status(HttpStatus.OK).body(this.userService.fetchAllUsers()) ;
     }
-    @PutMapping("/user")
-    public User getUpdateUserInfo(@RequestBody User userFromPostMan){
-        return this.userService.handleCreateUser(userFromPostMan);
+    @PutMapping("/users")
+    public ResponseEntity<User> getUpdateUserInfo(@RequestBody User userFromPostMan){
+        User fetchUpdateUser = this.userService.handleCreateUser(userFromPostMan);
+        return ResponseEntity.ok(fetchUpdateUser);
     }
 }
